Separate concerns of metric meter and recording
